pyproject.toml
[tool.poetry]
name = "blog.alexledger.net"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
PyYAML = "^6.0"
pydantic = "^1.10.5"
pandoc = "^2.3"
python-slugify = "^8.0.0"
notion-client = "^2.0.0"
Jinja2 = "^3.1.2"

[tool.poetry.dev-dependencies]
black = "^23.1.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
